requirements.txt
beautifulsoup4==4.13.4
colorama==0.4.6
Flask==3.1.1
flask_login==0.6.3
flask_sqlalchemy==3.1.1
websockets==15.0.1
Werkzeug==3.1.3
uvicorn==0.35.0
jwt==1.4.0
dotenv==0.9.9
fastapi==0.116.1
pydantic==2.11.7